Fix all 6 remaining build errors
1. MainWindow: Remove TitleBar.HeightOption (not in WindowsAppSDK 1.8)
2. MainWindow: Use AppWindow.MoveAndResize with Windows.Graphics.RectInt32
3. MarkdownRenderer: Use Windows.UI.Text.FontStyle (not Microsoft.UI.Text)
4. MarkdownRenderer: Use Windows.UI.Color.FromArgb (not Microsoft.UI.Color)
5. MemoryManager: Cast double to long for FromUnixTimeSeconds
6. SettingsPage: Define IInitializeWithWindow COM interface locally
   instead of relying on WinRT.InitializeWithWindow NuGet

+++ MemoryManager.cs
@@ -96,7 +96,7 @@ public static class MemoryManager
         };
         foreach (var e in entries)
         {
-            var when = DateTimeOffset.FromUnixTimeSeconds(e.Timestamp).DateTime.ToString("HH:mm:ss");
+            var when = DateTimeOffset.FromUnixTimeSeconds((long)e.Timestamp).DateTime.ToString("HH:mm:ss");
             var status = e.Success ? "OK" : "FAIL";
             var argsStr = string.Join(", ", e.Args.Select(kv => $"{kv.Key}={kv.Value}"));
             lines.Add($"- [{when}] {status} {e.ToolName}({argsStr}) -> {e.Result}");
